requirements.txt
# ============================================================
#       DRONE VISION AI - PYTHON REQUIREMENTS
# ============================================================
# Install with: pip install -r requirements.txt
# ============================================================

# ── DEEP LEARNING ────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ── OBJECT DETECTION (YOLOv8) ────────────────────────────────
ultralytics>=8.0.0

# ── FACE RECOGNITION ─────────────────────────────────────────
deepface>=0.0.79
retina-face>=0.0.13
tf-keras>=2.15.0

# ── OCR ──────────────────────────────────────────────────────
easyocr>=1.7.0

# ── COMPUTER VISION ──────────────────────────────────────────
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# ── DATA & MATH ──────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ── IMAGE PROCESSING ─────────────────────────────────────────
Pillow>=10.0.0

# ── DATABASE ─────────────────────────────────────────────────
sqlalchemy>=2.0.0

# ── CONFIG & UTILITIES ───────────────────────────────────────
pyyaml>=6.0
tqdm>=4.65.0

# ── SOUND ALERTS ─────────────────────────────────────────────
pygame>=2.5.0

# ── LOGGING & DISPLAY ────────────────────────────────────────
colorlog>=6.7.0
